+++ src/query/service/src/pipelines/processors/transforms/hash_join/common.rs
@@ -132,42 +132,6 @@ impl HashJoinProbeState {
         Ok(DataBlock::new_from_columns(vec![marker_column]))
     }
 
-    pub(crate) fn set_validity(
-        &self,
-        column: &BlockEntry,
-        num_rows: usize,
-        validity: &Bitmap,
-    ) -> BlockEntry {
-        let (value, data_type) = (&column.value, &column.data_type);
-        let col = value.convert_to_full_column(data_type, num_rows);
-
-        if matches!(col, Column::Null { .. }) {
-            column.clone()
-        } else if let Some(col) = col.as_nullable() {
-            if col.len() == 0 {
-                return BlockEntry::new(data_type.clone(), Value::Scalar(Scalar::Null));
-            }
-            // It's possible validity is longer than col.
-            let diff_len = validity.len() - col.validity.len();
-            let mut new_validity = MutableBitmap::with_capacity(validity.len());
-            for (b1, b2) in validity.iter().zip(col.validity.iter()) {
-                new_validity.push(b1 & b2);
-            }
-            new_validity.extend_constant(diff_len, false);
-            let col = Column::Nullable(Box::new(NullableColumn {
-                column: col.column.clone(),
-                validity: new_validity.into(),
-            }));
-            BlockEntry::new(data_type.clone(), Value::Column(col))
-        } else {
-            let col = Column::Nullable(Box::new(NullableColumn {
-                column: col.clone(),
-                validity: validity.clone(),
-            }));
-            BlockEntry::new(data_type.wrap_nullable(), Value::Column(col))
-        }
-    }
-
     // return an (option bitmap, all_true, all_false)
     pub(crate) fn get_other_filters(
         &self,
@@ -212,43 +176,6 @@ impl HashJoinProbeState {
                 column: other,
             }))),
         }
-    }
-
-    // Add `data_block` for build table to `row_space`
-    pub(crate) fn add_build_block(&self, data_block: DataBlock) -> Result<()> {
-        let mut data_block = data_block;
-        if matches!(
-            self.hash_join_state.hash_join_desc.join_type,
-            JoinType::Left | JoinType::LeftSingle | JoinType::Full
-        ) {
-            let mut validity = MutableBitmap::new();
-            validity.extend_constant(data_block.num_rows(), true);
-            let validity: Bitmap = validity.into();
-
-            let nullable_columns = data_block
-                .columns()
-                .iter()
-                .map(|c| self.set_validity(c, validity.len(), &validity))
-                .collect::<Vec<_>>();
-            data_block = DataBlock::new(nullable_columns, data_block.num_rows());
-        }
-
-        let chunk = Chunk { data_block };
-
-        {
-            // Acquire write lock in current scope
-            let mut chunks = self.hash_join_state.row_space.chunks.write();
-            if self.need_outer_scan() {
-                let outer_scan_map = unsafe { &mut *self.outer_scan_map.get() };
-                outer_scan_map.push(vec![false; chunk.num_rows()]);
-            }
-            if self.need_mark_scan() {
-                let mark_scan_map = unsafe { &mut *self.mark_scan_map.get() };
-                mark_scan_map.push(vec![MARKER_KIND_FALSE; chunk.num_rows()]);
-            }
-            chunks.push(chunk);
-        }
-        Ok(())
     }
 }
 
@@ -296,5 +223,41 @@ impl HashJoinState {
                 }
             }
         }
+    }
+}
+
+
+pub(crate) fn set_validity(
+    column: &BlockEntry,
+    num_rows: usize,
+    validity: &Bitmap,
+) -> BlockEntry {
+    let (value, data_type) = (&column.value, &column.data_type);
+    let col = value.convert_to_full_column(data_type, num_rows);
+
+    if matches!(col, Column::Null { .. }) {
+        column.clone()
+    } else if let Some(col) = col.as_nullable() {
+        if col.len() == 0 {
+            return BlockEntry::new(data_type.clone(), Value::Scalar(Scalar::Null));
+        }
+        // It's possible validity is longer than col.
+        let diff_len = validity.len() - col.validity.len();
+        let mut new_validity = MutableBitmap::with_capacity(validity.len());
+        for (b1, b2) in validity.iter().zip(col.validity.iter()) {
+            new_validity.push(b1 & b2);
+        }
+        new_validity.extend_constant(diff_len, false);
+        let col = Column::Nullable(Box::new(NullableColumn {
+            column: col.column.clone(),
+            validity: new_validity.into(),
+        }));
+        BlockEntry::new(data_type.clone(), Value::Column(col))
+    } else {
+        let col = Column::Nullable(Box::new(NullableColumn {
+            column: col.clone(),
+            validity: validity.clone(),
+        }));
+        BlockEntry::new(data_type.wrap_nullable(), Value::Column(col))
     }
 }
